% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pinging.R
\name{start_pinging}
\alias{start_pinging}
\title{Start continuous pinging}
\usage{
start_pinging(
  destination,
  time = 5,
  ip = "4.2.2.4",
  feedback = TRUE,
  timeUnit = "secs",
  interval = 1
)
}
\arguments{
\item{destination}{Path to the directory where the CSV log will be saved.}

\item{time}{Duration to run the pinging loop. Default is 5.}

\item{ip}{IP address to ping. Default is "4.2.2.4".}

\item{feedback}{If TRUE, prints each ping result to the console. Default is TRUE.}

\item{timeUnit}{Unit of time for the duration: "secs", "mins", "hours". Default is "secs".}

\item{interval}{Seconds between pings. Default is 1.}
}
\description{
Pings the given IP address repeatedly and logs results to a CSV file.
}
\examples{
\dontrun{
start_pinging("C:/Users/Desktop", time = 10, ip = "8.8.8.8", timeUnit = "secs")
}
}
